backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
cassandra-driver==3.29.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
